+++ src/utils/functions.ts
@@ -6,9 +6,11 @@ import {
   AngularDiameterMethod,
   AngularDistanceMethod,
   DifferentiableFunction,
+  EciVec3,
   JacobianFunction,
   Radians,
   SpaceObjectType,
+  Vec3,
 } from '../types/types';
 
 /**
@@ -299,6 +301,16 @@ export function angularDistance(
     default:
       throw new Error('Invalid angular distance method.');
   }
+}
+
+/**
+ * Calculates the linear distance between two points in three-dimensional space.
+ * @param pos1 The first position.
+ * @param pos2 The second position.
+ * @returns The linear distance between the two positions in kilometers.
+ */
+export function linearDistance<D extends number>(pos1: Vec3<D>, pos2: Vec3<D>): D {
+  return <D>Math.sqrt((pos1.x - pos2.x) ** 2 + (pos1.y - pos2.y) ** 2 + (pos1.z - pos2.z) ** 2);
 }
 
 /**
@@ -654,3 +666,34 @@ const spaceObjTypeStrMap = {
 
 export const spaceObjType2Str = (spaceObjType: SpaceObjectType): string =>
   spaceObjTypeStrMap[spaceObjType] || 'Unknown';
+
+export const dopplerFactor = (location: EciVec3, position: EciVec3, velocity: EciVec3): number => {
+  const mfactor = 7.292115e-5;
+  const c = 299792.458; // Speed of light in km/s
+
+  const range = <EciVec3>{
+    x: position.x - location.x,
+    y: position.y - location.y,
+    z: position.z - location.z,
+  };
+  const distance = Math.sqrt(range.x ** 2 + range.y ** 2 + range.z ** 2);
+
+  const rangeVel = <EciVec3>{
+    x: velocity.x + mfactor * location.y,
+    y: velocity.y - mfactor * location.x,
+    z: velocity.z,
+  };
+
+  const rangeRate = (range.x * rangeVel.x + range.y * rangeVel.y + range.z * rangeVel.z) / distance;
+  let dopplerFactor = 0;
+
+  if (rangeRate < 0) {
+    dopplerFactor = 1 + (rangeRate / c) * sign(rangeRate);
+  }
+
+  if (rangeRate >= 0) {
+    dopplerFactor = 1 - (rangeRate / c) * sign(rangeRate);
+  }
+
+  return dopplerFactor;
+};
